Decode/encode CBOR array to Go struct
This makes encoded data more compact and structs are easier to use.

	type T struct {
		_ struct{} `cbor:",toarray"`
		A int      `cbor:",omitempty"`
		B string   `cbor:",omitempty"`
	}
	v := T{A: 1, B: "hello"} // encoded to [1, "hello"]

Special field "_" is used to specify struct level options, such as
"toarray". Any value of T type is encoded as array of 2 elements.
"omitempty" is disabled by "toarray" to ensure that the same number
of elements are encoded every time.

If "toarray" is omitted, "omitempty" works just like encoding/json.

Closes: #30